# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "More-Recipes provides a platform for users to share the awesome and exciting recipe ideas they have invented or learnt.Suppose a user comes up with a recipe, he/she can post it on More-Recipes and get feedback in form of reviews and votes from other users who explore that recipe. Users can also keep a list of their favorite recipes on application.",
    "version": "1.0.0",
    "title": "More Recipes"
  },
  "host": "more-recipes-ovie.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/recipe": {
      "post": {
        "tags": [
          "Recipe"
        ],
        "summary": "Add recipe",
        "description": "Add recipe record.",
        "operationId": "addRecipe",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created recipe object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful recipe creation",
            "schema": {
              "$ref": "#/definitions/AddRecipe"
            }
          },
          "400": {
            "description": "Invalid Input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Recipe not found",
            "schema": {
              "$ref": "#/definitions/RecipeNotFound"
            }
          }
        }
      }
    },
    "/recipe/{recipeId}": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "summary": "Get recipe by recipeId",
        "description": "",
        "operationId": "getRecipeByRecipeId",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "The recipe that needs to be fetched.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe Retrieved SuccessFullly!",
            "schema": {
              "$ref": "#/definitions/GetRecipeById"
            }
          },
          "400": {
            "description": "Invalid recipeId"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Recipe Not Found!",
            "schema": {
              "$ref": "#/definitions/RecipeNotFound"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Recipe"
        ],
        "summary": "Update recipe",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateRecipe",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "recipe that needs to be updated",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated recipe object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe Upated SuccessFullly!",
            "schema": {
              "$ref": "#/definitions/UpdateRecipe"
            }
          },
          "400": {
            "description": "Invalid recipeId"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Recipe Not Found!",
            "schema": {
              "$ref": "#/definitions/RecipeNotFound"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recipe"
        ],
        "summary": "Delete recipe",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteRecipe",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "The recipe that needs to be deleted",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe Upated SuccessFullly!",
            "schema": {
              "$ref": "#/definitions/DeleteRecipe"
            }
          },
          "400": {
            "description": "Invalid recipeId"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Recipe Not Found!",
            "schema": {
              "$ref": "#/definitions/RecipeNotFound"
            }
          }
        }
      }
    },
...
    "/recipe/{recipeId}/review": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Post recipe review",
        "description": "Add review record for recipe.",
        "operationId": "postReview",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "Token Authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "recipe to post review on",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created review object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Review"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Review Posted Successfully",
            "schema": {
              "$ref": "#/definitions/PostRecipeReview"
            }
          },
          "400": {
            "description": "Invalid recipeId"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Recipe Not Found!",
            "schema": {
              "$ref": "#/definitions/RecipeNotFound"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Recipe": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "ingredients": {
          "type": "string"
        },
        "directions": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "views": {
          "type": "integer"
        },
        "upvotes": {
          "type": "integer"
        },
        "downvotes": {
          "type": "integer"
        },
        "notification": {
          "type": "integer",
          "items": {
            "$ref": "#/definitions/Review"
          }
        }
      }
    },
    "Review": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Review"
      }
    },
...
    "RecipeNotFound": {
      "type": "object",
      "required": [
        "status",
        "message"
      ],
      "properties": {
        "status": {
          "type": "string",
          "example": "fail"
        },
        "message": {
          "type": "string",
          "example": "Recipe Not Found!"
        }
      }
    },
    "AddRecipe": {
      "type": "object",
      "required": [
        "message",
        "recipe"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Recipe Added SuccessFullly!"
        },
        "recipe": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "ingredients": {
              "type": "string"
            },
            "directions": {
              "type": "string"
            },
            "imageUrl": {
              "type": "string"
            },
            "views": {
              "type": "integer"
            },
            "upvotes": {
              "type": "integer"
            },
            "downvotes": {
              "type": "integer"
            },
            "notification": {
              "type": "integer"
            }
          }
        }
      }
    },
    "UpdateRecipe": {
      "type": "object",
      "required": [
        "message",
        "recipe"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Recipe Upated SuccessFullly!"
        },
        "recipe": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "ingredients": {
              "type": "string"
            },
            "directions": {
              "type": "string"
            },
            "imageUrl": {
              "type": "string"
            },
            "views": {
              "type": "integer"
            },
            "upvotes": {
              "type": "integer"
            },
            "downvotes": {
              "type": "integer"
            },
            "notification": {
              "type": "integer"
            }
          }
        }
      }
    },
    "DeleteRecipe": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Recipe Deleted SuccessFullly!"
        }
      }
    },
    "GetRecipeById": {
      "type": "object",
      "required": [
        "message",
        "recipe"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Recipe Retrieved SuccessFullly!"
        },
        "recipe": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "ingredients": {
              "type": "string"
            },
            "directions": {
              "type": "string"
            },
            "imageUrl": {
              "type": "string"
            },
            "views": {
              "type": "integer"
            },
            "upvotes": {
              "type": "integer"
            },
            "downvotes": {
              "type": "integer"
            },
            "notification": {
              "type": "integer"
            }
          }
        }
      }
    },
...
    "PostRecipeReview": {
      "type": "object",
      "required": [
        "message",
        "review"
      ],
      "properties": {
        "message": {
          "type": "string",
          "example": "Review Posted Successfully"
        },
        "review": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "example": 1
            },
            "message": {
              "type": "string",
              "example": "nice meal"
            },
            "userId": {
              "type": "integer",
              "example": 1
            },
            "recipeId": {
              "type": "integer",
              "example": 1
            }
          }
        }
      }
    },
...
  },
...
}